chore: localize 파일 및 데이터 정리 	renamed:    src/libs/utils/__test__/localize.test.ts -> src/libs/utils/localize/__test__/localize.test.ts 	new file:   src/libs/utils/localize/localize.const.ts 	renamed:    src/libs/utils/localize.ts -> src/libs/utils/localize/localize.ts 	renamed:    src/libs/types/localize.type.ts -> src/libs/utils/localize/localize.type.ts 	modified:   src/services/httpClient/httpClient.type.ts 	modified:   src/services/httpClient/utils.ts 	modified:   src/services/openWeatherOneCall/axios.ts 	modified:   src/services/openWeatherOneCall/mock/mock.ts 	modified:   src/services/openWeatherOneCall/status.ts
