+++ src/libs/utils/localize/localize.ts
@@ -1,13 +1,6 @@
-import type { LanguageCode, LocalizedTextMap } from '@libs/types/localize.type';
+import { DEFAULT_LANGUAGE, GET_LOCALIZED_TEXT_FROM_MAP_ERROR } from '@libs/utils/localize/localize.const';
 
-export const DEFAULT_LANGUAGE: LanguageCode = 'kr';
-
-export const GET_LOCALIZED_TEXT_FROM_MAP_ERROR: LocalizedTextMap<'10001'> = {
-  10001: {
-    en: 'Invalid key in localizedTextMap.',
-    kr: 'localizedTextMap에 존재하지 않는 키입니다.',
-  },
-};
+import type { LanguageCode, LocalizedTextMap } from '@libs/utils/localize/localize.type';
 
 /**
  * 로컬라이징된 텍스트를 반환하는 함수
